+++ src/spring/HISTORY.md
@@ -2,7 +2,7 @@ Release History
 ===============
 1.5.0
 ---
-* Add argument `--client-auth-certificates` in command `az spring app create` and `az spring app update`.
+* Add argument `--client-auth-certs` in command `az spring app create` and `az spring app update`.
 
 1.4.0
 ---